+++ handlers/get_tasks_id.go
@@ -1,23 +1,24 @@
-// Obtener una tarea por su ID
 package handlers
 
 import (
+	"database/sql"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
 
-func GetTasksId() {
-	r.GET("/tasks/:id", func(c *gin.Context) {
-		id := c.Param("id")
+func GetTasksId(c *gin.Context) {
 
-		var task Task
-		err := db.QueryRow("SELECT id, title, description, due_date, completed FROM tasks WHERE id=$1", id).Scan(&task.ID, &task.Title, &task.Description, &task.DueDate, &task.Completed)
-		if err != nil {
-			c.AbortWithError(http.StatusNotFound, err)
-			return
-		}
+	db, err := sql.Open("postgres", "postgres://postgres:mysecretpassword@localhost:5432/postgres?sslmode=disable")
 
-		c.JSON(http.StatusOK, task)
-	})
+	id := c.Param("id")
+
+	var task Task
+	err = db.QueryRow("SELECT id, title, description, due_date, completed FROM tasks WHERE id=$1", id).Scan(&task.ID, &task.Title, &task.Description, &task.DueDate, &task.Completed)
+	if err != nil {
+		c.AbortWithError(http.StatusNotFound, err)
+		return
+	}
+
+	c.JSON(http.StatusOK, task)
 }
