+++ handlers/get_tasks_pending.go
@@ -1,31 +1,32 @@
-// Obtener todas las tareas no completadas
 package handlers
 
 import (
+	"database/sql"
 	"net/http"
 
 	"github.com/gin-gonic/gin"
 )
 
-func GetTasksPending() {
-	r.GET("/tasks/pending", func(c *gin.Context) {
-		rows, err := db.Query("SELECT id, title, description, due_date, completed FROM tasks WHERE completed=false ORDER BY id")
-		if err != nil {
+func GetTasksPending(c *gin.Context) {
+
+	db, err := sql.Open("postgres", "postgres://postgres:mysecretpassword@localhost:5432/postgres?sslmode=disable")
+
+	rows, err := db.Query("SELECT id, title, description, due_date, completed FROM tasks WHERE completed=false ORDER BY id")
+	if err != nil {
+		c.AbortWithError(http.StatusInternalServerError, err)
+		return
+	}
+	defer rows.Close()
+
+	tasks := []Task{}
+	for rows.Next() {
+		var task Task
+		if err := rows.Scan(&task.ID, &task.Title, &task.Description, &task.DueDate, &task.Completed); err != nil {
 			c.AbortWithError(http.StatusInternalServerError, err)
 			return
 		}
-		defer rows.Close()
+		tasks = append(tasks, task)
+	}
 
-		tasks := []Task{}
-		for rows.Next() {
-			var task Task
-			if err := rows.Scan(&task.ID, &task.Title, &task.Description, &task.DueDate, &task.Completed); err != nil {
-				c.AbortWithError(http.StatusInternalServerError, err)
-				return
-			}
-			tasks = append(tasks, task)
-		}
-
-		c.JSON(http.StatusOK, tasks)
-	})
+	c.JSON(http.StatusOK, tasks)
 }
